Route /repo/:owner/:repo/contents/:path to the .md file
This PR uses the ReactRouter's HashRouter to route /repo/:owner/:repo/contents/:path to the .md file in GitHub instead of using query args.

This will make it easier to add a different component to handle routing of URLs such as:
/repo/:owner/:repo

At this URL for example, we can provide a list of the .md files available in the repo that the user can click on.