Overhaul of GitHub repository configuration
This change provides a mechanism to configure repositories in Music Markdown.

We also overhauled how we store GitHub repos in localStorage. Repos are now treated as a string of the form `repo-owner/repo-name` instead of as an object. This should make Jesse happy as he wanted this in the first place.

To support this, I updated the routing rules to treat `repo-owner/repo-name` as a single token, which required going through all the code and replacing references of (owner, repo) pairs with just a single reference.

I also added tests for the github requests. This involved introducing a mock for the global fetch.