Add a GitHub token dialog box.
The dialog box allows you to add a GitHub Token to Music Markdown which
enables you to create or update Music Markdown files in a GitHub
repository.

If a GitHub Token isn't present, we disable the add button in the file
browser, and the save button in the file editor. We don't disable the
edit button since editing doesn't require a token.