Fix eval of slowScroll JS.
Recent updates patches a vulnerability in dangerouslySetInnerHTML that
can cause XSS exploit but breaks slowScroll of guitar chord tab. This
hack fixes slowScroll, but re-introduces the XSS vulnerability. I've
added a TODO to come back and use an IFrame to fix this permanently,
but that fix is going to require reworking of markdown-it-music.